atomic: correct dependencies on atomic
bug: v/63628
